+++ client/src/components/MarkerIcons/groomerIcon.ts
