+++ client/src/components/MarkerIcons/userPositionIcon.ts
