+++ client/src/components/MarkerIcons/vetIcon.ts
